Suggest adding notes on choice of set of controls and creating codebook
Three sets of control variable exists. There is no explanation on why one set was chosen and deemed better than the other. Furthermore, this is hard to trace out because the variables that differ across the three sets lack variable labels as well as a codebook. I suggest adding notes and providing a codebook to enhance reproducibility.